requirements.txt
amqp==5.0.6
appnope==0.1.2
asgiref==3.4.1
backcall==0.2.0
billiard==3.6.4.0
celery==5.2.1
certifi==2021.10.8
charset-normalizer==2.0.8
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
decorator==5.1.0
Deprecated==1.2.13
Django==3.2.9
django-celery-beat==2.2.1
django-extensions==3.1.5
django-queryset-csv==1.1.0
django-redis==5.0.0
django-rest-swagger==2.2.0
django-timezone-field==4.2.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
h11==0.12.0
idna==3.3
ipython==7.30.0
itypes==1.2.0
jedi==0.18.1
Jinja2==3.0.3
kombu==5.2.2
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
numpy==1.21.4
openapi-codec==1.3.2
pandas==1.3.4
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.23
psycopg2-binary==2.9.2
ptyprocess==0.7.0
Pygments==2.10.0
PyJWT==2.3.0
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2021.3
redis==4.0.2
requests==2.26.0
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
traitlets==5.1.1
unicodecsv==0.14.1
uritemplate==4.1.1
urllib3==1.26.7
uvicorn==0.15.0
vine==5.0.0
wcwidth==0.2.5
wrapt==1.13.3
